Add nds_schema.py for a clear code
